pyproject.toml
[tool.poetry]
name = "sso"
version = "0.1.0"
description = ""
authors = ["ya_novikov <arhdomain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
redis = "^6.0.0"
pyjwt = "^2.10.1"
alembic = "^1.15.2"
pydantic-settings = "^2.9.1"
passlib = "^1.7.4"
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"
black = "^25.1.0"
mypy = "^1.15.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
fakeredis = "^2.28.1"
pytest-asyncio = "^0.26.0"


[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
python_classes = "*Test*"
python_files = "tests.py test_*.py"
pythonpath = "src"
testpaths = ["tests"]

[tool.mypy]


[tool.ruff]
lint.ignore = ["E712"]
lint.select = [
  "E", # pycodestyle
  "F", # Pyflakes
  "UP", # pyupgrade
  "B", # flake8-bugbear
  "SIM", # flake8-simplify
  "I", # isort
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = [
  "A",
  "B",
  "C",
  "D",
  "E",
  "F",
  "G",
  "I",
  "N",
  "Q",
  "S",
  "T",
  "W",
  "ANN",
  "ARG",
  "BLE",
  "COM",
  "DJ",
  "DTZ",
  "EM",
  "ERA",
  "EXE",
  "FBT",
  "ICN",
  "INP",
  "ISC",
  "NPY",
  "PD",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "RET",
  "RSE",
  "RUF",
  "SIM",
  "SLF",
  "TCH",
  "TID",
  "TRY",
  "UP",
  "YTT",
]
lint.unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
  "*migrations*",
  "app/app/management/commands/*",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py311"

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 18

[tool.ruff.lint.isort]
lines-after-imports = 2
lines-between-types = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
